requirements.txt
absl-py==1.3.0
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
contourpy==1.0.5
cvzone==1.5.6
cycler==0.11.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.37.4
kiwisolver==1.4.4
matplotlib==3.6.1
mediapipe==0.8.11
numpy==1.23.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
